requirements.txt
openpyxl~=3.1.5
ezdxf~=1.3.3
shapely~=2.0.6
python-docx~=1.1.2
pandas~=2.2.3